Move reporting file into reporting sub-folder
