+++ casper/src/test/scala/coop/rchain/casper/batch1/MultiParentCasperReportingSpec.scala
@@ -5,7 +5,7 @@ import coop.rchain.casper.helper.TestNode.Effect
 import coop.rchain.casper.protocol.CommEvent
 import coop.rchain.casper.rholang.Resources
 import coop.rchain.casper.util.ConstructDeploy
-import coop.rchain.casper.{ReportStore, ReportingCasper}
+import coop.rchain.casper.reporting.{ReportStore, ReportingCasper}
 import coop.rchain.models.{BindPattern, ListParWithRandom, Par, TaggedContinuation}
 import coop.rchain.p2p.EffectsTestInstances.LogicalTime
 import coop.rchain.rspace.ReportingRspace.ReportingComm
