+++ node/src/test/scala/coop/rchain/node/TransactionAPISpec.scala
@@ -5,7 +5,7 @@ import coop.rchain.casper.helper.TestNode
 import coop.rchain.casper.rholang.Resources
 import coop.rchain.casper.util.ConstructDeploy
 import coop.rchain.casper.util.GenesisBuilder.{buildGenesis, GenesisContext}
-import coop.rchain.casper.{ReportStore, ReportingCasper}
+import coop.rchain.casper.reporting.{ReportStore, ReportingCasper}
 import coop.rchain.crypto.PrivateKey
 import coop.rchain.crypto.signatures.Secp256k1
 import coop.rchain.models.Par
